requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiodns==3.4.0
    # via aiohttp
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.11
    # via lead-generation
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
beautifulsoup4==4.13.4
    # via lead-generation
brotli==1.1.0 ; platform_python_implementation == 'CPython'
    # via aiohttp
brotlicffi==1.1.0.0 ; platform_python_implementation != 'CPython'
    # via aiohttp
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   brotlicffi
    #   pycares
    #   zstandard
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.12
    # via lead-generation
fastapi-cli==0.0.7
    # via fastapi
frozenlist==1.6.2
    # via
    #   aiohttp
    #   aiosignal
google-search-results==2.4.2
    # via lead-generation
greenlet==3.2.3 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via fastapi
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via lead-generation
langchain-core==0.3.64
    # via
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.21
    # via lead-generation
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.45
    # via
    #   langchain
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
openai==1.85.0
    # via langchain-openai
orjson==3.10.18 ; platform_python_implementation != 'PyPy'
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pycares==4.8.0
    # via aiodns
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via
    #   lead-generation
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
    #   uvicorn
regex==2024.11.6
    # via tiktoken
requests==2.32.4
    # via
    #   google-search-results
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.7
    # via fastapi-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.41
    # via langchain
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via langchain-core
tiktoken==0.9.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typer==0.16.0
    # via fastapi-cli
typing-extensions==4.14.0
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn==0.34.3
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
